+++ tests/test_015_resultset.rs
@@ -9,14 +9,13 @@ use serde_derive::Deserialize;
 #[test] // cargo test --test test_015_resultset -- --nocapture
 pub fn test_015_resultset() -> HdbResult<()> {
     let mut log_handle = test_utils::init_logger();
+    let start = std::time::Instant::now();
     let mut connection = test_utils::get_authenticated_connection()?;
 
     evaluate_resultset(&mut log_handle, &mut connection)?;
     verify_row_ordering(&mut log_handle, &mut connection)?;
 
-    info!("{} calls to DB were executed", connection.get_call_count()?);
-
-    Ok(())
+    test_utils::closing_info(connection, start)
 }
 
 fn evaluate_resultset(