+++ tests/test_016_selectforupdate.rs
@@ -9,14 +9,13 @@ use std::time::Duration;
 #[test] // cargo test --test test_016_selectforupdate -- --nocapture
 pub fn test_016_selectforupdate() -> HdbResult<()> {
     let mut log_handle = test_utils::init_logger();
+    let start = std::time::Instant::now();
     let mut connection = test_utils::get_authenticated_connection()?;
 
     prepare(&mut log_handle, &mut connection)?;
     produce_conflicts(&mut log_handle, &mut connection)?;
 
-    info!("{} calls to DB were executed", connection.get_call_count()?);
-
-    Ok(())
+    test_utils::closing_info(connection, start)
 }
 
 fn prepare(_log_handle: &mut ReconfigurationHandle, connection: &mut Connection) -> HdbResult<()> {