+++ tests/test_027_small_decimals.rs
@@ -12,12 +12,12 @@ use serde_derive::Deserialize;
 #[test]
 fn test_027_small_decimals() -> HdbResult<()> {
     let mut log_handle = test_utils::init_logger();
+    let start = std::time::Instant::now();
     let mut connection = test_utils::get_authenticated_connection()?;
 
     test_small_decimals(&mut log_handle, &mut connection)?;
 
-    info!("{} calls to DB were executed", connection.get_call_count()?);
-    Ok(())
+    test_utils::closing_info(connection, start)
 }
 
 fn test_small_decimals(