+++ tests/test_028_timestamp.rs
@@ -8,12 +8,12 @@ use log::{debug, info, trace};
 #[test]
 pub fn test_028_timestamp() -> HdbResult<()> {
     let mut log_handle = test_utils::init_logger();
+    let start = std::time::Instant::now();
     let mut connection = test_utils::get_authenticated_connection()?;
 
     test_timestamp(&mut log_handle, &mut connection)?;
 
-    info!("{} calls to DB were executed", connection.get_call_count()?);
-    Ok(())
+    test_utils::closing_info(connection, start)
 }
 
 // Test the conversion of timestamps