+++ tests/test_033_clobs.rs
@@ -12,14 +12,14 @@ use std::io::Read;
 #[test]
 pub fn test_033_clobs() -> HdbResult<()> {
     let mut log_handle = test_utils::init_logger();
+    let start = std::time::Instant::now();
     let mut connection = test_utils::get_authenticated_connection()?;
 
     let (blabla, fingerprint) = get_blabla();
     test_clobs(&mut log_handle, &mut connection, &blabla, &fingerprint)?;
     test_streaming(&mut log_handle, &mut connection, blabla, &fingerprint)?;
 
-    info!("{} calls to DB were executed", connection.get_call_count()?);
-    Ok(())
+    test_utils::closing_info(connection, start)
 }
 
 fn get_blabla() -> (String, Vec<u8>) {