+++ tests/test_060_metadata.rs
@@ -7,12 +7,12 @@ use log::{debug, info};
 #[test]
 pub fn test_060_metadata() -> HdbResult<()> {
     let mut log_handle = test_utils::init_logger();
+    let start = std::time::Instant::now();
     let mut connection = test_utils::get_authenticated_connection()?;
 
     test_procedure_metadata(&mut log_handle, &mut connection)?;
 
-    info!("{} calls to DB were executed", connection.get_call_count()?);
-    Ok(())
+    test_utils::closing_info(connection, start)
 }
 
 fn test_procedure_metadata(